RenderSystem scale now works, changed Scale from int to float

+++ src/Components/RenderComponent.cs
@@ -5,9 +5,9 @@ namespace Kazaam.Objects {
   public class RenderComponent {
     public Texture2D Texture {get; set;}  
     public SpriteEffects Effects {get; set;}
-    public int Scale {get; set;}
+    public float Scale {get; set;}
 
-    public RenderComponent(Sprite sprite, bool flip = false, int scale = 1) {
+    public RenderComponent(Sprite sprite, bool flip = false, float scale = 1.0f) {
       Texture = sprite.Texture;
       Effects = flip ? SpriteEffects.FlipHorizontally : SpriteEffects.None;
       Scale = scale;
